backend/requirements.txt
fastapi
asyncpg
pydantic

python-jose
bcrypt
nanoid
